refactor: revise Like balance of wallet
