+++ src/components/GQL/queries/walletBalance.ts
@@ -12,6 +12,10 @@ export default gql`
           id
         }
       }
+      liker {
+        total
+        rateUSD
+      }
     }
   }
 `
